Fixed some warnings on x86_64 Linux
git-svn-id: https://hypertable.googlecode.com/svn/trunk@467 01261837-d133-0410-8dcb-6792eb736156

+++ src/cc/Common/StringExt.h
@@ -76,7 +76,7 @@ inline std::string operator+( const std::string& s1, long long llval ) {
 
 inline std::string operator+( const std::string& s1, uint64_t llval ) {
   char cbuf[32];
-  sprintf(cbuf, "%lld", llval);
+  sprintf(cbuf, "%llu", (long long unsigned int)llval);
   return s1 + cbuf;
 }
